+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 # Changelog
 
-All notable changes to `:package_name` will be documented in this file.
+All notable changes to `nano-converter-php` will be documented in this file.
 
-## 1.0.0 - 202X-XX-XX
+## 1.0.0 - 2021-04-12
 
-- initial release
+- Add NANO class to convert NANO to Raw
+- Add Raw class to convert Raw to NANO